+++ modules/os2forms_attachment/src/Os2formsAttachmentPrintBuilder.php
@@ -3,13 +3,27 @@
 namespace Drupal\os2forms_attachment;
 
 use Drupal\Core\Entity\EntityInterface;
+use Drupal\Core\File\FileExists;
+use Drupal\Core\File\FileSystemInterface;
+use Drupal\Core\StringTranslation\TranslationInterface;
+use Drupal\entity_print\Event\PreSendPrintEvent;
+use Drupal\entity_print\Event\PrintEvents;
 use Drupal\entity_print\Plugin\PrintEngineInterface;
 use Drupal\entity_print\PrintBuilder;
+use Drupal\entity_print\Renderer\RendererFactoryInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * The OS2Forms attachment print builder service.
  */
 class Os2formsAttachmentPrintBuilder extends PrintBuilder {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function __construct(RendererFactoryInterface $renderer_factory, EventDispatcherInterface $event_dispatcher, TranslationInterface $string_translation, protected readonly FileSystemInterface $file_system) {
+    parent::__construct($renderer_factory, $event_dispatcher, $string_translation);
+  }
 
   /**
    * {@inheritdoc}
@@ -28,9 +42,55 @@ class Os2formsAttachmentPrintBuilder extends PrintBuilder {
   }
 
   /**
+   * Modified version of the original savePrintable() function.
+   *
+   * The only difference is modified call to prepareRenderer with digitalPost
+   * flag TRUE.
+   *
+   * @see PrintBuilder::savePrintable()
+   *
+   * @return string
+   *   FALSE or the URI to the file. E.g. public://my-file.pdf.
+   */
+  public function savePrintableDigitalSignature(array $entities, PrintEngineInterface $print_engine, $scheme = 'public', $filename = FALSE, $use_default_css = TRUE) {
+    $renderer = $this->prepareRenderer($entities, $print_engine, $use_default_css, TRUE);
+
+    // Allow other modules to alter the generated Print object.
+    $this->dispatcher->dispatch(new PreSendPrintEvent($print_engine, $entities), PrintEvents::PRE_SEND);
+
+    // If we didn't have a URI passed in the generate one.
+    if (!$filename) {
+      $filename = $renderer->getFilename($entities) . '.' . $print_engine->getExportType()->getFileExtension();
+    }
+
+    $uri = "$scheme://$filename";
+
+    // Save the file.
+    return $this->file_system->saveData($print_engine->getBlob(), $uri, FileExists::Replace);
+  }
+
+  /**
    * {@inheritdoc}
    */
-  protected function prepareRenderer(array $entities, PrintEngineInterface $print_engine, $use_default_css) {
+
+  /**
+   * Override prepareRenderer() the print engine with the passed entities.
+   *
+   * @param array $entities
+   *   An array of entities.
+   * @param \Drupal\entity_print\Plugin\PrintEngineInterface $print_engine
+   *   The print engine.
+   * @param bool $use_default_css
+   *   TRUE if we want the default CSS included.
+   * @param bool $digitalSignature
+   *   If the digital signature message needs to be added.
+   *
+   * @return \Drupal\entity_print\Renderer\RendererInterface
+   *   A print renderer.
+   *
+   * @see PrintBuilder::prepareRenderer
+   */
+  protected function prepareRenderer(array $entities, PrintEngineInterface $print_engine, $use_default_css, $digitalSignature = FALSE) {
     if (empty($entities)) {
       throw new \InvalidArgumentException('You must pass at least 1 entity');
     }
@@ -50,6 +110,9 @@ class Os2formsAttachmentPrintBuilder extends PrintBuilder {
     // structure. That margin is automatically added in PDF and PDF only.
     $generatedHtml = (string) $renderer->generateHtml($entities, $render, $use_default_css, TRUE);
     $generatedHtml .= "<style>fieldset legend {margin-left: -12px;}</style>";
+    if ($digitalSignature) {
+      $generatedHtml .= $this->t('You can validate the signature on this PDF file via validering.nemlog-in.dk.');
+    }
 
     $print_engine->addPage($generatedHtml);