+++ modules/os2forms_fasit/src/Form/SettingsForm.php
@@ -2,39 +2,63 @@
 
 namespace Drupal\os2forms_fasit\Form;
 
-use Drupal\Core\Form\FormBase;
+use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Form\ConfigFormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\os2forms_fasit\Helper\CertificateLocatorHelper;
-use Drupal\os2forms_fasit\Helper\Settings;
+use Drupal\os2forms_fasit\Helper\FasitHelper;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\OptionsResolver\Exception\ExceptionInterface as OptionsResolverException;
 
 /**
- * Organisation settings form.
+ * Fasit settings form.
  */
-final class SettingsForm extends FormBase {
+final class SettingsForm extends ConfigFormBase {
   use StringTranslationTrait;
 
+  public const CONFIG_NAME = 'os2forms_fasit.settings';
   public const FASIT_API_BASE_URL = 'fasit_api_base_url';
   public const FASIT_API_TENANT = 'fasit_api_tenant';
   public const FASIT_API_VERSION = 'fasit_api_version';
   public const CERTIFICATE = 'certificate';
+  public const KEY = 'key';
+  public const CERTIFICATE_PROVIDER = 'certificate_provider';
+  public const PROVIDER_TYPE_FORM = 'form';
+  public const PROVIDER_TYPE_KEY = 'key';
 
-  /**
-   * Constructor.
-   */
-  public function __construct(private readonly Settings $settings, private readonly CertificateLocatorHelper $certificateLocatorHelper) {
-  }
+  public const ACTION_PING_API = 'action_ping_api';
 
   /**
    * {@inheritdoc}
    */
-  public static function create(ContainerInterface $container): SettingsForm {
+  public function __construct(
+    ConfigFactoryInterface $config_factory,
+    private readonly FasitHelper $helper,
+  ) {
+    parent::__construct($config_factory);
+  }
+
+  /**
+   * {@inheritdoc}
+   *
+   * @phpstan-return self
+   */
+  public static function create(ContainerInterface $container): self {
     return new static(
-      $container->get(Settings::class),
-      $container->get(CertificateLocatorHelper::class)
+      $container->get('config.factory'),
+      $container->get(FasitHelper::class)
     );
+  }
+
+  /**
+   * {@inheritdoc}
+   *
+   * @phpstan-return array<string>
+   */
+  protected function getEditableConfigNames() {
+    return [
+      self::CONFIG_NAME,
+    ];
   }
 
   /**
@@ -51,57 +75,74 @@ final class SettingsForm extends FormBase {
    * @phpstan-return array<string, mixed>
    */
   public function buildForm(array $form, FormStateInterface $form_state): array {
+    $form = parent::buildForm($form, $form_state);
+    $config = $this->config(self::CONFIG_NAME);
 
-    $fasitApiBaseUrl = $this->settings->getFasitApiBaseUrl();
     $form[self::FASIT_API_BASE_URL] = [
       '#type' => 'textfield',
       '#title' => $this->t('Fasit API base url'),
       '#required' => TRUE,
-      '#default_value' => !empty($fasitApiBaseUrl) ? $fasitApiBaseUrl : NULL,
+      '#default_value' => $config->get(self::FASIT_API_BASE_URL),
       '#description' => $this->t('Specifies which base url to use. This is disclosed by Schultz'),
     ];
 
-    $fasitApiTenant = $this->settings->getFasitApiTenant();
     $form[self::FASIT_API_TENANT] = [
       '#type' => 'textfield',
       '#title' => $this->t('Fasit API tenant'),
       '#required' => TRUE,
-      '#default_value' => !empty($fasitApiTenant) ? $fasitApiTenant : NULL,
+      '#default_value' => $config->get(self::FASIT_API_TENANT),
       '#description' => $this->t('Specifies which tenant to use. This is disclosed by Schultz'),
     ];
 
-    $fasitApiVersion = $this->settings->getFasitApiVersion();
     $form[self::FASIT_API_VERSION] = [
       '#type' => 'textfield',
       '#title' => $this->t('Fasit API version'),
       '#required' => TRUE,
-      '#default_value' => !empty($fasitApiVersion) ? $fasitApiVersion : NULL,
+      '#default_value' => $config->get(self::FASIT_API_VERSION),
       '#description' => $this->t('Specifies which api version to use. Should probably be v2'),
     ];
 
-    $certificate = $this->settings->getCertificate();
+    $certificateConfig = $config->get(self::CERTIFICATE) ?? [];
 
     $form[self::CERTIFICATE] = [
       '#type' => 'fieldset',
       '#title' => $this->t('Certificate'),
       '#tree' => TRUE,
 
-      CertificateLocatorHelper::LOCATOR_TYPE => [
+      self::CERTIFICATE_PROVIDER => [
         '#type' => 'select',
-        '#title' => $this->t('Certificate locator type'),
+        '#title' => $this->t('Provider'),
         '#options' => [
-          CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT => $this->t('Azure key vault'),
-          CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM => $this->t('File system'),
+          self::PROVIDER_TYPE_FORM => $this->t('Form'),
+          self::PROVIDER_TYPE_KEY => $this->t('Key'),
         ],
-        '#default_value' => $certificate[CertificateLocatorHelper::LOCATOR_TYPE] ?? NULL,
+        '#default_value' => $certificateConfig[self::CERTIFICATE_PROVIDER] ?? self::PROVIDER_TYPE_FORM,
+        '#description' => $this->t('Specifies which provider to use'),
       ],
+    ];
+
+    $form[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_TYPE] = [
+      '#type' => 'select',
+      '#title' => $this->t('Certificate locator type'),
+      '#options' => [
+        CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT => $this->t('Azure key vault'),
+        CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM => $this->t('File system'),
+      ],
+      '#default_value' => $certificateConfig[CertificateLocatorHelper::LOCATOR_TYPE] ?? NULL,
+      '#states' => [
+        'visible' => [':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_FORM]],
+      ],
+      '#description' => $this->t('Specifies which locator to use'),
     ];
 
     $form[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT] = [
       '#type' => 'fieldset',
       '#title' => $this->t('Azure key vault'),
       '#states' => [
-        'visible' => [':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT]],
+        'visible' => [
+          ':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_FORM],
+          ':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT],
+        ],
       ],
     ];
 
@@ -118,9 +159,12 @@ final class SettingsForm extends FormBase {
       $form[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT][$key] = [
         '#type' => 'textfield',
         '#title' => $info['title'],
-        '#default_value' => $certificate[CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT][$key] ?? NULL,
+        '#default_value' => $certificateConfig[CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT][$key] ?? NULL,
         '#states' => [
-          'required' => [':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT]],
+          'required' => [
+            ':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_FORM],
+            ':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_AZURE_KEY_VAULT],
+          ],
         ],
       ];
     }
@@ -129,15 +173,21 @@ final class SettingsForm extends FormBase {
       '#type' => 'fieldset',
       '#title' => $this->t('File system'),
       '#states' => [
-        'visible' => [':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM]],
+        'visible' => [
+          ':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_FORM],
+          ':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM],
+        ],
       ],
 
       'path' => [
         '#type' => 'textfield',
         '#title' => $this->t('Path'),
-        '#default_value' => $certificate[CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM]['path'] ?? NULL,
+        '#default_value' => $certificateConfig[CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM]['path'] ?? NULL,
         '#states' => [
-          'required' => [':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM]],
+          'required' => [
+            ':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_FORM],
+            ':input[name="certificate[locator_type]"]' => ['value' => CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM],
+          ],
         ],
       ],
     ];
@@ -145,20 +195,36 @@ final class SettingsForm extends FormBase {
     $form[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_PASSPHRASE] = [
       '#type' => 'textfield',
       '#title' => $this->t('Passphrase'),
-      '#default_value' => $certificate[CertificateLocatorHelper::LOCATOR_PASSPHRASE] ?? NULL,
+      '#default_value' => $certificateConfig[CertificateLocatorHelper::LOCATOR_PASSPHRASE] ?? NULL,
+      '#states' => [
+        'visible' => [
+          ':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_FORM],
+        ],
+      ],
     ];
 
-    $form['actions']['#type'] = 'actions';
-
-    $form['actions']['submit'] = [
-      '#type' => 'submit',
-      '#value' => $this->t('Save settings'),
+    $form[self::CERTIFICATE][self::PROVIDER_TYPE_KEY] = [
+      '#type' => 'key_select',
+      '#title' => $this->t('Key'),
+      '#default_value' => $certificateConfig[self::PROVIDER_TYPE_KEY] ?? NULL,
+      '#states' => [
+        'visible' => [':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_KEY]],
+        'required' => [':input[name="certificate[certificate_provider]"]' => ['value' => self::PROVIDER_TYPE_KEY]],
+      ],
     ];
 
-    $form['actions']['testCertificate'] = [
-      '#type' => 'submit',
-      '#name' => 'testCertificate',
-      '#value' => $this->t('Test certificate'),
+    $form['actions']['ping_api'] = [
+      '#type' => 'container',
+
+      self::ACTION_PING_API => [
+        '#type' => 'submit',
+        '#name' => self::ACTION_PING_API,
+        '#value' => $this->t('Ping API'),
+      ],
+
+      'message' => [
+        '#markup' => $this->t('Note: Pinging the API will use saved config.'),
+      ],
     ];
 
     return $form;
@@ -169,20 +235,23 @@ final class SettingsForm extends FormBase {
    *
    * @phpstan-param array<string, mixed> $form
    */
-  public function validateForm(array &$form, FormStateInterface $formState): void {
-    $triggeringElement = $formState->getTriggeringElement();
-    if ('testCertificate' === ($triggeringElement['#name'] ?? NULL)) {
+  public function validateForm(array &$form, FormStateInterface $form_state): void {
+    if (self::ACTION_PING_API === ($form_state->getTriggeringElement()['#name'] ?? NULL)) {
       return;
     }
 
-    $values = $formState->getValues();
+    $values = $form_state->getValues();
 
-    if (CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM === $values[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_TYPE]) {
-      $path = $values[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM]['path'] ?? NULL;
-      if (!file_exists($path)) {
-        $formState->setErrorByName('certificate][file_system][path', $this->t('Invalid certificate path: %path', ['%path' => $path]));
+    if (self::PROVIDER_TYPE_FORM === $values[self::CERTIFICATE][self::CERTIFICATE_PROVIDER]) {
+      if (CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM === $values[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_TYPE]) {
+        $path = $values[self::CERTIFICATE][CertificateLocatorHelper::LOCATOR_TYPE_FILE_SYSTEM]['path'] ?? NULL;
+        if (!file_exists($path)) {
+          $form_state->setErrorByName('certificate][file_system][path', $this->t('Invalid certificate path: %path', ['%path' => $path]));
+        }
       }
     }
+
+    parent::validateForm($form, $form_state);
   }
 
   /**
@@ -190,44 +259,30 @@ final class SettingsForm extends FormBase {
    *
    * @phpstan-param array<string, mixed> $form
    */
-  public function submitForm(array &$form, FormStateInterface $formState): void {
-    $triggeringElement = $formState->getTriggeringElement();
-    if ('testCertificate' === ($triggeringElement['#name'] ?? NULL)) {
-      $this->testCertificate();
+  public function submitForm(array &$form, FormStateInterface $form_state): void {
+    if (self::ACTION_PING_API === ($form_state->getTriggeringElement()['#name'] ?? NULL)) {
+      try {
+        $this->helper->pingApi();
+        $this->messenger()->addStatus($this->t('Pinged API successfully.'));
+      }
+      catch (\Throwable $t) {
+        $this->messenger()->addError($this->t('Pinging API failed: @message', ['@message' => $t->getMessage()]));
+      }
       return;
     }
 
-    try {
-      $settings[self::CERTIFICATE] = $formState->getValue(self::CERTIFICATE);
-      $settings[self::FASIT_API_BASE_URL] = $formState->getValue(self::FASIT_API_BASE_URL);
-      $settings[self::FASIT_API_TENANT] = $formState->getValue(self::FASIT_API_TENANT);
-      $settings[self::FASIT_API_VERSION] = $formState->getValue(self::FASIT_API_VERSION);
-
-      $this->settings->setSettings($settings);
-      $this->messenger()->addStatus($this->t('Settings saved'));
+    $config = $this->config(self::CONFIG_NAME);
+    foreach ([
+      self::FASIT_API_BASE_URL,
+      self::FASIT_API_TENANT,
+      self::FASIT_API_VERSION,
+      self::CERTIFICATE,
+    ] as $key) {
+      $config->set($key, $form_state->getValue($key));
     }
-    catch (OptionsResolverException $exception) {
-      $this->messenger()->addError($this->t('Settings not saved (@message)', ['@message' => $exception->getMessage()]));
+    $config->save();
 
-      return;
-    }
-
-    $this->messenger()->addStatus($this->t('Settings saved'));
-  }
-
-  /**
-   * Test certificate.
-   */
-  private function testCertificate(): void {
-    try {
-      $certificateLocator = $this->certificateLocatorHelper->getCertificateLocator();
-      $certificateLocator->getCertificates();
-      $this->messenger()->addStatus($this->t('Certificate successfully tested'));
-    }
-    catch (\Throwable $throwable) {
-      $message = $this->t('Error testing certificate: %message', ['%message' => $throwable->getMessage()]);
-      $this->messenger()->addError($message);
-    }
+    parent::submitForm($form, $form_state);
   }
 
 }